YH-985: move ManageProfile to widgets layer, separate setActiveProfile, createProfile, deleteProfile features
