+++ public/locales/ru/profile.json
@@ -1,84 +1,85 @@
 {
-  "edit.page.title": "Редактирование профиля",
-  "tabs": {
-    "personal": "Личная информация",
-    "about.me": "Обо мне",
-    "skills": "Навыки"
-  },
-  "photo": {
-    "title": "Фото профиля",
-    "description": "Ваше фото будет видно всем членам сообщества YeaHub",
-    "update": "Обновить фото",
-    "delete": "Удалить фото",
-    "update.full": "Обновить фотографию",
-    "delete.full": "Удалить фотографию",
-    "modal": {
-      "title": "Выберите область для фотографий профиля",
-      "description": "Выбранная миниатюра будет использоваться в статьях, личных сообщениях и комментариях.",
-      "submit": "Сохраните изменения",
-      "click.secondary": "Измените фото",
-      "large.preview": "Большой размер превью аватара",
-      "small.preview": "Маленький размер превью аватара",
-      "min.res": "Минимальное разрешение для изображения: {{minRes}}",
-      "max.res": "Максимальное разрешение для изображения: {{maxRes}}"
-    }
-  },
-  "personal": {
-    "title": "Личная информация",
-    "description": "Ваша подробная информация"
-  },
-  "links": {
-    "title": "Личные ссылки",
-    "description": "Поделитесь своими профилями в других соц. сетях"
-  },
-  "about.me": {
-    "title": "О себе любимом(-ой)",
-    "description": "Расскажи о себе всему сообществу. Мы ценим человека не за его профессиональные качества,поэтому пиши всё чем хочешь поделиться",
-    "empty": "Нет описания"
-  },
-  "skills": {
-    "title": "Твои навыки",
-    "description": "Покажи, что ты умеешь и в чём ты действительно хорош",
-    "empty": "Не выбраны навыки"
-  },
-  "form": {
-    "username": "Никнейм",
-    "specialization": "IT специальность",
-    "email": "Email для связи",
-    "location": "Локация",
-    "skills": "Навыки"
-  },
-  "settings.tabs": {
-    "select.tariff": "Выбрать тариф",
-    "change.password": "Смена пароля",
-    "verify.email": "Подтверждение e-mail",
+	"edit.page.title": "Редактирование профиля",
+	"tabs": {
+		"personal": "Личная информация",
+		"about.me": "Обо мне",
+		"skills": "Навыки"
+	},
+	"photo": {
+		"title": "Фото профиля",
+		"description": "Ваше фото будет видно всем членам сообщества YeaHub",
+		"update": "Обновить фото",
+		"delete": "Удалить фото",
+		"update.full": "Обновить фотографию",
+		"delete.full": "Удалить фотографию",
+		"modal": {
+			"title": "Выберите область для фотографий профиля",
+			"description": "Выбранная миниатюра будет использоваться в статьях, личных сообщениях и комментариях.",
+			"submit": "Сохраните изменения",
+			"click.secondary": "Измените фото",
+			"large.preview": "Большой размер превью аватара",
+			"small.preview": "Маленький размер превью аватара",
+			"min.res": "Минимальное разрешение для изображения: {{minRes}}",
+			"max.res": "Максимальное разрешение для изображения: {{maxRes}}"
+		}
+	},
+	"personal": {
+		"title": "Личная информация",
+		"description": "Ваша подробная информация"
+	},
+	"links": {
+		"title": "Личные ссылки",
+		"description": "Поделитесь своими профилями в других соц. сетях"
+	},
+	"about.me": {
+		"title": "О себе любимом(-ой)",
+		"description": "Расскажи о себе всему сообществу. Мы ценим человека не за его профессиональные качества,поэтому пиши всё чем хочешь поделиться",
+		"empty": "Нет описания"
+	},
+	"skills": {
+		"title": "Твои навыки",
+		"description": "Покажи, что ты умеешь и в чём ты действительно хорош",
+		"empty": "Не выбраны навыки"
+	},
+	"form": {
+		"username": "Никнейм",
+		"specialization": "IT специальность",
+		"email": "Email для связи",
+		"location": "Локация",
+		"skills": "Навыки"
+	},
+	"settings.tabs": {
+		"select.tariff": "Выбрать тариф",
+		"change.password": "Смена пароля",
+		"verify.email": "Подтверждение e-mail",
 		"account": "Аккаунт"
-  },
-  "email.verification": {
-    "title": "Подтвердите ваш e-mail",
-    "description": "Подтвердите ваш e-mail для активации учетной записи, после активации вы получите доступ к нашей платформе",
-    "email": "Введите e-mail",
-    "submit": "Подтвердить",
-    "sent.success": "Письмо успешно отправлено на почту",
-    "verify.stub": {
-      "title": "{{username}}, подтвердите ваш e-mail",
-      "description": "Станьте частью нашей команды! Подтвердите ваш e-mail, чтобы получить доступ к дополнительным возможностям платформы и начать расширять свои горизонты в сообществе таких же целеустремленных профессионалов, как и вы. Присоединяйтесь к нам на пути к новым вершинам!",
-      "link": "Подтвердить e-mail"
-    },
-    "success.verify.stub": {
-      "title": "Почта успешно подтверждена",
-      "description": "Спасибо что подтвердили вашу почту, теперь вам доступно больше услуг на платформе"
-    }
-  },
-  "change.password": {
-    "title": "Сменить пароль",
-    "description": "В целях безопасности или если вы заметили подозрительную активность смените пароль",
-    "label": "Введите новый пароль",
-    "repeat.label": "Повторите пароль",
-    "placeholder": "Пароль"
-  },
+	},
+	"email.verification": {
+		"title": "Подтвердите ваш e-mail",
+		"description": "Подтвердите ваш e-mail для активации учетной записи, после активации вы получите доступ к нашей платформе",
+		"email": "Введите e-mail",
+		"submit": "Подтвердить",
+		"sent.success": "Письмо успешно отправлено на почту",
+		"verify.stub": {
+			"title": "{{username}}, подтвердите ваш e-mail",
+			"description": "Станьте частью нашей команды! Подтвердите ваш e-mail, чтобы получить доступ к дополнительным возможностям платформы и начать расширять свои горизонты в сообществе таких же целеустремленных профессионалов, как и вы. Присоединяйтесь к нам на пути к новым вершинам!",
+			"link": "Подтвердить e-mail"
+		},
+		"success.verify.stub": {
+			"title": "Почта успешно подтверждена",
+			"description": "Спасибо что подтвердили вашу почту, теперь вам доступно больше услуг на платформе"
+		}
+	},
+	"change.password": {
+		"title": "Сменить пароль",
+		"description": "В целях безопасности или если вы заметили подозрительную активность смените пароль",
+		"label": "Введите новый пароль",
+		"repeat.label": "Повторите пароль",
+		"placeholder": "Пароль"
+	},
 	"manage.profiles": {
 		"title": "Ваши профили",
+		"open": "Перейти",
 		"tooltip": {
 			"not.member": "Дополнительные профили доступны членам сообщества",
 			"max.count": "Вы достигли лимита создаваемых профилей"
