+++ src/features/profile/createProfile/index.ts
@@ -1,0 +1,1 @@
+export { CreateProfileButton } from './ui/CreateProfileButton/CreateProfileButton';
