+++ src/features/profile/createProfile/model/constants/createProfileConstants.ts
@@ -1,0 +1,3 @@
+export const createProfileApiUrls = {
+	createProfile: '/profiles',
+};
