+++ src/features/profile/createProfile/model/types/createProfileTypes.ts
@@ -1,0 +1,8 @@
+import { Profile } from '@/entities/auth';
+
+export type CreateProfileRequestData = Pick<
+	Profile,
+	'profileType' | 'specializationId' | 'markingWeight'
+> & {
+	userId: string;
+};
