+++ src/features/profile/deleteProfile/index.ts
@@ -1,0 +1,1 @@
+export { DeleteProfileButton } from './ui/DeleteProfileButton/DeleteProfileButton';
