+++ src/features/profile/deleteProfile/model/constants/deleteProfileContants.ts
@@ -1,0 +1,3 @@
+export const deleteProfileApiUrls = {
+	deleteProfile: '/profiles/:profileId',
+};
