+++ src/features/profile/deleteProfile/model/types/deleteProfileTypes.ts
@@ -1,0 +1,5 @@
+import { ModalProps } from '@/shared/ui/Modal';
+
+export type DeleteProfileModalProps = Pick<ModalProps, 'isOpen' | 'onClose'> & {
+	profileId: string;
+};
