--- src/features/profile/manageProfiles/model/constants/manageProfilesContants.ts
@@ -1,1 +1,0 @@
-export const profilesCountLimit = 10;
