--- src/features/profile/manageProfiles/ui/ManageProfilesPanel/ManageProfilesPanel.module.css
@@ -1,19 +1,0 @@
-.container {
-  padding: 24px;
-}
-
-.container > *  {
-  margin-bottom: 30px;
-}
-
-.title {
-  line-height: 127%;
-}
-
-.mb {
-  margin-bottom: 30px;
-}
-
-.create-button {
-  align-self: flex-end;
-}
