--- src/features/profile/manageProfiles/ui/ProfileCard/ProfileCard.module.css
@@ -1,69 +1,0 @@
-.container {
-  position: relative;
-}
-
-.image {
-  max-width: 38px;
-  max-height: 38px;
-  border: 1px solid var(--color-black-30);
-  border-radius: 38px;
-}
-
-.title {
-  position: relative;
-  color: var(--color-blue-550);
-  line-height: 120%;
-}
-
-.card {
-  flex-grow: 1;
-  padding: 10px 24px;
-  border: 1px solid var(--color-black-30);
-  border-radius: 14px;
-  background-color: var(--color-black-30);
-  cursor: pointer;
-
-  &:hover {
-    border: 1px solid var(--color-purple-700);
-
-    .title {
-      color: var(--color-purple-700);
-    }
-
-    .image {
-      border: 1px solid var(--color-purple-700);
-    }
-  }
-}
-
-.icon-check::after {
-  position: absolute;
-  bottom: -10px;
-  left: -26px;
-  display: block;
-  width: 18px;
-  height: 18px;
-  background-image: url('../../model/assets/check.avif');
-  background-position: center;
-  background-size: contain;
-  background-repeat: no-repeat;
-  content: '';
-}
-
-.active-container {
-  cursor: none;
-}
-
-.active-color {
-  color: var(--color-purple-700);
-}
-
-.active-border {
-  border: 1px solid var(--color-purple-700);
-}
-
-.delete {
-  position: absolute;
-  top: 20px;
-  right: 24px;
-}
