+++ src/features/profile/setActiveProfile/index.ts
@@ -1,0 +1,1 @@
+export { useSetActiveProfileMutation } from './api/setActiveProfileApi';
