+++ src/features/profile/setActiveProfile/model/constants/setActiveProfileConstants.ts
@@ -1,0 +1,3 @@
+export const setActiveProfileApiUrls = {
+	setActiveProfile: '/profiles/:profileId/active',
+};
