+++ src/widgets/manageProfiles/model/assets/index.ts
