+++ src/widgets/manageProfiles/model/constants/manageProfilesContants.ts
@@ -1,0 +1,2 @@
+export const NOT_MEMBER_PROFILES_COUNT_LIMIT = 1;
+export const MEMBER_PROFILES_COUNT_LIMIT = 10;
