+++ src/widgets/manageProfiles/ui/ManageProfilesHeader/ManageProfilesHeader.module.css
@@ -1,0 +1,3 @@
+.title {
+  line-height: 127%;
+}
