+++ src/widgets/manageProfiles/ui/ManageProfilesPanel/ManageProfilesPanel.module.css
@@ -1,0 +1,11 @@
+.container {
+  padding: 24px;
+}
+
+.mb {
+  margin-bottom: 30px;
+}
+
+.create-button {
+  align-self: flex-end;
+}
